requirements.txt
Flask
Flask-SocketIO
sounddevice
geocoder
numpy
tensorflow
librosa
PyYAML
scikit-learn